Fixed bugs in generalized input reading.
[Bugs]
- Fixed some minor bugs/syntax problem with the generalization of the
  input reading. The input reading now works well enough on its own, but
  still needs to be incorporated with the rest of the system.

[File's changed]
	modified:   input/__init__.py
	modified:   input/gzoltar_input.py
	new file:   input/input_reader.py
	modified:   input/tcm_input.py
